+++ NEWS.md
@@ -7,6 +7,7 @@
 * add use_data_*() functions
 * drop use of `fct_explicit_na()` (deprecated) (#15)
 * create_data use_data return datasets as tibble 
+* allowed to run explore() without data 
 
 ## explore 1.0.2 (2023-01-14)